+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Better Combat support [#694]
 - A new creative tab with sample gear items, mostly intended to aid with testing
+### Changed
+- Furnace fuel burn times are now handled by the NeoForge data map; netherwood charcoal burn time config option removed
 ### Fixed
 - Alloy makers missing their last input slot [#705]
+- Cannot use netherwood sticks as fuel [#672]
 - All sickle behavior errors [#703]
 - Sinew and flax mapping to the wrong material [#700]
 - Netherwood sticks becoming regular sticks during tool crafting